+++ app/src/reducers/storyReducer.js
@@ -1,3 +1,8 @@
+
+import { Story } from '../utils/DataFormat.js'
+import { INTENT_URL, UTTER_URL } from '../utils/url_routes.js';
+import axios from "axios";
+
 const INITIAL_STATE = {
     intents: [],
     utters: [],
@@ -5,52 +10,8 @@ const INITIAL_STATE = {
     selected_utter_position: -1,
     id_item: "",
     old_item: {},
-    items: [
-        {
-            "id": 1,
-            "name": "story_basoioiica",
-            "formatted_content": [
-                {
-                    "id": 1,
-                    "type": "intent",
-                    "name": "cumprimentar"
-                },
-                {
-                    "id": 2,
-                    "type": "utter",
-                    "name": "utter_cumprimentar"
-                }
-            ]
-        }
-    ],
-    current_item: {
-        "id": 1,
-        "name": "story_basica",
-        "formatted_content": [
-            {
-                "id": 1,
-                "type": "intent",
-                "name": "cumprimentar"
-            },
-            {
-                "id": 2,
-                "type": "utter",
-                "name": "utter_cumprimentar"
-            }
-        ]
-    },
-    item_contents: [
-        {
-            "id": 1,
-            "type": "intent",
-            "name": "cumprimentar"
-        },
-        {
-            "id": 2,
-            "type": "utter",
-            "name": "utter_cumprimentar"
-        }
-    ],
+    items: [],
+    item_contents: [],
     old_item_contents: [],
     name_item: "",
     old_name_item: "",
@@ -76,34 +37,16 @@ export default (state = INITIAL_STATE, action) => {
     switch (action.type) {
 
         case "GET_INTENTS": {
-
-            let selected_item_position = 0;
-
-            let selected_item = state.intents.find((item, index) => {
-                selected_item_position = index;
-                return (item.id === action.item.id || item.name === action.item.name);
-            });
-
             return {
                 ...state,
                 intents: action.intents,
-                selected_intents_position: selected_item_position
             }
         }
 
         case "GET_UTTERS": {
-
-            let selected_item_position = 0;
-
-            let selected_item = state.utters.find((item, index) => {
-                selected_item_position = index;
-                return (item.id === action.utters.id || item.name === action.intents.name);
-            });
-
             return {
                 ...state,
                 utters: action.utters,
-                selected_utters_position: selected_item_position
             }
         }
 
@@ -143,4 +86,84 @@ export default (state = INITIAL_STATE, action) => {
             return state;
         }
     }
+};
+
+export const getIntents = () => {
+    return async (dispatch) => {
+        try {
+            const response = await axios.get(INTENT_URL);
+            await dispatch({ type: "GET_INTENTS", intents: response.data });
+        } catch (error) {
+            throw (error);
+        }
+    }
+};
+
+export const getUtters = () => {
+    return async (dispatch) => {
+        try {
+            const response = await axios.get(UTTER_URL);
+            await dispatch({ type: "GET_UTTERS", utters: response.data });
+        } catch (error) {
+            throw (error);
+        }
+    }
+};
+
+export const reorderList = (arr) =>{
+
+    return {
+        type: "REORDER",
+        items: arr
+    }
+    
+}
+
+export const addToStory = (item) => {
+
+    return {
+        type: "ADD_TO_STORY",
+        item: item
+    }
+}
+
+export const removeItem = (item_position) => {
+    return {
+        type: "REMOVE_ITEM",
+        item_position: item_position
+    }
+}
+
+
+export const createOrUpdateItem = (mode = 'post', url = "", new_item, message = "") => {
+    return async (dispatch) => {
+        try {
+            const mode_url = (mode === 'post') ? url : url + new_item.id;
+
+            await axios[mode](mode_url, new_item);
+            dispatch(notifyAction(message));
+
+        } catch (error) {
+            throw (error);
+        }
+    }
+};
+
+export const saveData = (url, item) => {
+    console.log("SAASDAS", item)
+    return async (dispatch) => {
+        if ((item.id === "" || item.id===undefined)) {
+            dispatch(createOrUpdateItem('post', url, item, "Story criada com sucesso!"));
+        } else {
+            dispatch(createOrUpdateItem('put', url, item, "Story atualizada com sucesso!"));
+        }
+    }
+}
+
+
+export const notifyAction = (text) => {
+    return {
+        type: "SUCESS_ACTION_ITEM",
+        text: text
+    };
 };
